requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
fire==0.4.0
idna==3.3
requests==2.27.1
six==1.16.0
termcolor==1.1.0
urllib3==1.26.7
python-dotenv~=0.19.2